refactor: Extract audio file processing logic into a dedicated `AudioFileProcessor` class
This commit refactors the `EkaScribe` facade by moving the complex audio file processing and transcription pipeline logic into a new internal `AudioFileProcessor` class. This improves maintainability and separates concerns within the SDK.

- **New `AudioFileProcessor` Class:**
    - Created `AudioFileProcessor.kt` to handle the end-to-end workflow for pre-recorded files: session initialization, audio chunking (25s segments), S3 uploading, and transaction lifecycle management (init, stop, commit, poll).
    - Encapsulated logic for uploading both chunked segments and the full original recording.

- **`EkaScribe` Refactoring:**
    - Removed redundant processing logic and private helper methods (like `getAudioDurationMs`) from the `EkaScribe` class.
    - Updated `processAudioFile` to delegate execution to the new `AudioFileProcessor`.
    - Simplified SDK initialization and resource cleanup by managing a single `AudioFileProcessor` instance instead of multiple scattered references (`chunkUploaderRef`, `encoderRef`, etc.).
